agente-financiero-guias-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic[email]
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
cryptography
git+https://github.com/reingart/pyafipws.git@py3k
PySimpleSOAP
httplib2
reportlab
httpx
pytest
pytest-asyncio
python-multipart
fernet
google-cloud-storage
python-dateutil
